+++ symulation_components/depot/states.py
@@ -1,0 +1,9 @@
+from enum import Enum
+
+
+class StopState(Enum):
+    """
+    Utility enum representing Stop State
+    """
+    EMPTY = 0
+    BUSY = 1
